Re-organization and cleaning of the tutorial directory
