+++ .idea/.name
@@ -1,1 +1,1 @@
-procedural-finger.py
+gripper_simulation.py
