--- docs/sofapython3/tutorials/ProceduralFinger/.idea/.gitignore
@@ -1,3 +1,0 @@
-# Default ignored files
-/shelf/
-/workspace.xml
