--- docs/sofapython3/tutorials/ProceduralFinger/.idea/.name
@@ -1,1 +1,0 @@
-SoftFingerDesign
