+++ docs/sofapython3/tutorials/SoftFingerDesign/details/Data/SG90_servomotor_finger.stl
