+++ docs/sofapython3/tutorials/SoftFingerDesign/details/Data/cylinder.stl
