+++ docs/sofapython3/tutorials/SoftFingerDesign/details/Data/finger.stl
