+++ docs/sofapython3/tutorials/SoftFingerDesign/details/Data/finger_surface_contact_in2.stl
