+++ docs/sofapython3/tutorials/SoftFingerDesign/details/Data/servo_collision.stl
