+++ docs/sofapython3/tutorials/SoftFingerDesign/details/cylinder.py
