+++ docs/sofapython3/tutorials/SoftFingerDesign/details/finger_cura.3mf
