+++ docs/sofapython3/tutorials/SoftFingerDesign/details/gripper_simulation.py
