+++ docs/sofapython3/tutorials/SoftFingerDesign/images/assembly.png
