+++ docs/sofapython3/tutorials/SoftFingerDesign/images/fingers.png
