+++ docs/sofapython3/tutorials/SoftFingerDesign/images/parameters.png
