+++ docs/sofapython3/tutorials/SoftFingerDesign/images/printing_building_plate.PNG
