+++ docs/sofapython3/tutorials/SoftFingerDesign/images/proceduraltuto.xcf
