+++ docs/sofapython3/tutorials/SoftFingerDesign/images/simulation.png
