+++ docs/sofapython3/tutorials/SoftFingerDesign/images/softfingerdesign_header.png
