+++ docs/sofapython3/tutorials/SoftFingerDesign/soft-finger-design.py
@@ -1,0 +1,2 @@
+def createScene(rootNode):
+    return rootNode
